Fixed BetterRolls active tab resetting after every change
